requirements.txt
tensorflow>=2.9.0
wfdb>=3.4.1
scikit-learn>=1.0.2
numpy>=1.21.0
matplotlib>=3.5.1
scipy>=1.7.3
pyserial>=3.5
